Corrected the rewiring, drafted path length, modularity.
Hard to explain in words, but the mindist and maxdist nodes should have been selected from the sets of unconnected and connected edges, respectively. This wasn't the case for the code last week, and corrected that, and ran 4 brains with different eps parameters (0.2, 0.3, 0.4, 0.5).

Diameter and average path length are implemented in igraph, modularity needs decision about clustering. Should clean that and put it in the for loop of _plot.R script.